+++ lib/oneview-sdk/resource/api2000/c7000/server_hardware_type.rb
@@ -9,12 +9,16 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-require_relative '../../api1800/c7000/lig_uplink_set'
+require_relative '../../api1800/c7000/server_hardware_type'
 
 module OneviewSDK
   module API2000
     module C7000
-      class LIGUplinkSet < OneviewSDK::API1800::C7000::LIGUplinkSet
+      # Server hardware type resource implementation for API2000 C7000
+      class ServerHardwareType < OneviewSDK::API1800::C7000::ServerHardwareType
+        def self.api_version
+          2000
+        end
       end
     end
   end