+++ lib/oneview-sdk/resource/api2000/synergy/server_hardware_type.rb
@@ -9,12 +9,13 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-require_relative '../../api1800/c7000/lig_uplink_set'
+require_relative '../../api1800/synergy/server_hardware_type'
 
 module OneviewSDK
   module API2000
-    module C7000
-      class LIGUplinkSet < OneviewSDK::API1800::C7000::LIGUplinkSet
+    module Synergy
+      # Server hardware type resource implementation for API2000 Synergy
+      class ServerHardwareType < OneviewSDK::API1800::Synergy::ServerHardwareType
       end
     end
   end