+++ spec/unit/resource/api2000/c7000/lig_uplink_set.rb
@@ -9,13 +9,12 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-require_relative '../../api1800/c7000/lig_uplink_set'
+require 'spec_helper'
 
-module OneviewSDK
-  module API2000
-    module C7000
-      class LIGUplinkSet < OneviewSDK::API1800::C7000::LIGUplinkSet
-      end
-    end
+RSpec.describe OneviewSDK::API2000::C7000::LIGUplinkSet do
+  include_context 'shared context'
+
+  it 'inherits from OneviewSDK::API1800::C7000::LIGUplinkSet' do
+    expect(described_class).to be < OneviewSDK::API1800::C7000::LIGUplinkSet
   end
 end